+++ city.h
@@ -14,16 +14,16 @@ public:
     std::string countryName;
     double lattitude;
     double longitude;
-    int timeZoneGMTOffset;
+    double timeZoneGMTOffset;
 
     //Default Constructor
     City();
 
     // Extended Constructor
-    City(int p_cityId, std::string p_cityName, std::string p_countryName, double p_lattitude, double p_longitude, int p_timeZoneGMTOffset );
+    City(int p_cityId, std::string p_cityName, std::string p_countryName, double p_lattitude, double p_longitude, double p_timeZoneGMTOffset );
 
     // Easy method to quickly set all properties
-    void   setAllParameters(int p_cityId, std::string p_cityName, std::string p_countryName, double p_lattitude, double p_longitude, int p_timeZoneGMTOffset);
+    void   setAllParameters(int p_cityId, std::string p_cityName, std::string p_countryName, double p_lattitude, double p_longitude, double p_timeZoneGMTOffset);
 
     // String return of key values in city object
     std::string debugString();
